1.修复有些情况下获取考勤组成员失败的问题 2.修复【自动获取Token】未勾选时保存设置报错。 3.修复一些重名问题

+++ ali_dingtalk/models/res_config_settings.py
@@ -82,5 +82,5 @@ class ResConfigSettings(models.TransientModel):
         else:
             cron = self.env['ir.cron'].sudo().search(
                 [('code', '=', "env['ali.dingtalk.get.token'].get_token()")])
-            cron.sudo().unlink()
-
+            if len(cron) >= 1:
+                cron.sudo().unlink()
